Change journeymen and reservation controllers to function according to authorization

+++ app/controllers/v1/reservations_controller.rb
@@ -2,9 +2,17 @@ class V1::ReservationsController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    reservations = Reservation.all
+    reservations = if current_user.admin?
+      Reservation.all
+    else
+      current_user.reservations
+    end
 
-    render json: ReservationSerializer.new(reservations).serializable_hash[:data], status: :ok
+    if reservations == []
+      render json: { message: 'No reservations found' }, status: :not_found
+    else
+      render json: ReservationSerializer.new(reservations).serializable_hash[:data], status: :ok
+    end
   end
 
   def show
@@ -29,13 +37,16 @@ class V1::ReservationsController < ApplicationController
   end
 
   def destroy
-    reservation = Reservation.find_by(id: params[:id])
-
-    if reservation.nil?
-      render status: 404, json: { error: 'Reservation not found' }.to_json
+    if current_user.admin?
+      reservation = Reservation.find_by(id: params[:id])
+      if reservation.nil?
+        render status: 404, json: { error: 'Reservation not found' }.to_json
+      else
+        reservation.destroy
+        render json: { message: 'Reservation deleted' }.to_json
+      end
     else
-      reservation.destroy
-      render json: { message: 'Reservation deleted' }.to_json
+      render json: { message: 'You are not authorized to perform this action' }, status: :unauthorized
     end
   end
